requirements.txt
# requirements.txt

asyncgui==0.6.3
asynckivy==0.6.4
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
docutils==0.21.2
exceptiongroup==1.2.2
idna==3.10
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd2==2.0.1
materialyoucolor==2.0.9
pillow==11.0.0
Pygments==2.18.0
pypiwin32==223
pywin32==308
requests==2.32.3
urllib3==2.2.3
